+++ Configs/ItemActions.cs
@@ -1,5 +1,7 @@
 using System.ComponentModel;
+using BetterInventory.ItemActions;
 using SpikysLib.Configs;
+using Terraria.ID;
 using Terraria.ModLoader.Config;
 
 namespace BetterInventory.Configs;
@@ -11,16 +13,21 @@ public sealed class ItemActions : ModConfig {
     [DefaultValue(true)] public bool favoritedBuff;
     [DefaultValue(true)] public bool builderAccs;
     [DefaultValue(true)] public bool keepSwappedFavorited;
+    public Toggle<ShowBagContent> showBagContent = new(true);
 
     public static bool FastContainerOpening => Instance.fastContainerOpening;
     public static bool FastExtractinator => Instance.fastExtractinator;
     public static bool FavoritedBuff => Instance.favoritedBuff;
     public static bool BuilderAccs => Instance.builderAccs;
     public static bool KeepSwappedFavorited => Instance.keepSwappedFavorited;
+    public static bool ShowBagContent => Instance.showBagContent;
 
     public override ConfigScope Mode => ConfigScope.ClientSide;
     public static ItemActions Instance = null!;
 
+    public sealed override void OnChanged() {
+        BetterPlayer.GetGrabBagContent(ItemID.None);
+    }
 }
 
 public sealed class ItemRightClick {
@@ -28,4 +35,11 @@ public sealed class ItemRightClick {
 
     public static bool Enabled => ItemActions.Instance.itemRightClick;
     public static ItemRightClick Value => ItemActions.Instance.itemRightClick.Value;
+}
+
+public sealed class ShowBagContent {
+    public bool compact;
+    
+    public static bool Enabled => ItemActions.Instance.showBagContent;
+    public static ShowBagContent Value => ItemActions.Instance.showBagContent.Value;
 }