+++ Changelog.md
@@ -4,6 +4,7 @@
 - Added an options to keep swapped items favorited
 - Added the number of visible recipes to the recipe list
 - Prevented the offset of the recipes when at the bottom of the recipe list
+- Added grab bag drops in their tooltip
 
 ### v0.6.1.3
 - Fixed guideTile saving as lens when empty